# Zadanie REST/src/main/java/sk/stuba/fei/uim/vsa/pr2/resources/CreateThesisRequest.java
package sk.stuba.fei.uim.vsa.pr2.resources;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class CreateThesisRequest {
    private String registrationNumber;
    private String title;
    private String description;
    private String type;
}